feat(readme file, top evel): description updated on top level ReadMe file
The top level ReadMe file was edited to include some more information about the general features of
the application. The information includes all of the different features that are planned to be
delieverd on the project. It also includes the Notion documentation, which was used to originally
introduce the project to developers. A couple of weird typos were also fixed